deleted now irrelevant files
